chore: improve developing experience on local
