+++ .npmrc
@@ -130,7 +130,6 @@ link-workspace-packages=false # deep
 ## script to handle this.
 ##
 sync-injected-deps-after-scripts[]=build:pkg
-sync-injected-deps-after-scripts[]=build:infra
 sync-injected-deps-after-scripts[]=build:glint
 sync-injected-deps-after-scripts[]=sync
