+++ internal-tooling/src/sync-licence.ts
@@ -1,0 +1,4 @@
+#! /usr/bin/env bun
+import { main } from './tasks/sync-license';
+
+main();
