+++ packages/codemods/package.json
@@ -14,8 +14,7 @@
   },
   "scripts": {
     "build:cli": "bun build bin/codemods.ts --compile --target node --outfile codemods --sourcemap",
-    "lint": "eslint . --quiet --cache --cache-strategy=content",
-    "_syncPnpm": "bun run sync-dependencies-meta-injected"
+    "lint": "eslint . --quiet --cache --cache-strategy=content"
   },
   "exports": {
     ".": {
