+++ packages/ember/package.json
@@ -19,7 +19,9 @@
     "build:glint": "glint && glint --build",
     "build:pkg": "vite build;",
     "prepack": "pnpm run build:pkg",
-    "sync": "echo \"syncing\""
+    "sync": "echo \"syncing\"",
+    "lint": "eslint . --quiet --cache --cache-strategy=content",
+    "start": "vite"
   },
   "files": [
     "unstable-preview-types",
