+++ packages/legacy-compat/package.json
@@ -42,7 +42,8 @@
     "lint": "eslint . --quiet --cache --cache-strategy=content",
     "build:pkg": "vite build;",
     "prepack": "pnpm run build:pkg",
-    "sync": "echo \"syncing\""
+    "sync": "echo \"syncing\"",
+    "start": "vite"
   },
   "ember-addon": {
     "main": "addon-main.cjs",
