+++ tests/ember-data__adapter/package.json
@@ -16,11 +16,13 @@
   },
   "scripts": {
     "build:tests": "IS_TESTING=true EMBER_CLI_TEST_COMMAND=true ember build --output-path=dist-test --suppress-sizes",
-    "_build:production": "pnpm build:tests -e production",
+    "build:production": "bun run build:tests -e production",
     "lint": "eslint . --quiet --cache --cache-strategy=content",
     "check:types": "tsc --noEmit",
     "test": "bun ./diagnostic.js",
-    "_test:production": "bun ./diagnostic.js"
+    "test:production": "bun ./diagnostic.js",
+    "start": "bun run build:tests --watch",
+    "test:start": "bun ./diagnostic.js --serve --no-launch"
   },
   "dependencies": {
     "@babel/core": "^7.26.9",
