+++ tests/main/package.json
@@ -14,14 +14,15 @@
   },
   "scripts": {
     "build:tests": "IS_TESTING=true EMBER_CLI_TEST_COMMAND=true ember build --output-path=dist-test --suppress-sizes",
-    "build:production": "pnpm build:tests -e production",
+    "build:production": "bun run build:tests -e production",
     "lint": "eslint . --quiet --cache --cache-strategy=content",
     "check:types": "glint",
     "examine": "export EXAM_PARALLEL_COUNT=$(./bin/calculate-test-jobs); ember exam --test-port=0 --path=dist-test --parallel=$EXAM_PARALLEL_COUNT --load-balance",
     "launch:tests": "ember test --test-port=0 --serve --no-launch",
     "start": "bun run build:tests --watch",
     "test": "bun run examine",
-    "test:production": "bun run examine"
+    "test:production": "bun run examine",
+    "test:start": "bun run ember test --test-port=0 --path=dist-test --serve --no-launch"
   },
   "author": "",
   "license": "MIT",
